Feito o exercício 3
